[#100] Centralized save_chat_history on user_to_client fn
