+++ backend/tests/routes/test_chat_routes.py
@@ -56,9 +56,10 @@ def test_get_chats_authenticated(client: TestClient, session: Session) -> None:
     last_chat_message = content["chats"][1].get("last_message")
     assert last_chat_message is not None
     sender = last_chat_message.get("sender_role")
-    assert sender == Sender_Role_Enum.USER.value
-    assert last_chat_message.get("message") == "Saved message with image"
-    assert last_chat_message.get("media")
+    assert sender == Sender_Role_Enum.CLIENT.value
+    assert (
+        last_chat_message.get("message") == "Yes, I need help with something."
+    )
 
 
 def test_get_chat_details_by_client_id(
